changelog, doc, renaming and update of lumi install
