+++ cli/lra/aqua_lra.j2
@@ -7,14 +7,16 @@
 #SBATCH --time={{ time }}
 #SBATCH --mem={{ memory }}
 #SBATCH --error={{ log_error }}
-#SBATCH --output={{ log_error }}
-{% if dependency != None %}
+#SBATCH --output={{ log_output }}
+{% if dependency %}
 #SBATCH --dependency={{ dependency }}
 {% endif %}
 
 
 {% if singularity %}
+{% if machine == 'lumi' %}
 unset SINGULARITY_BIND  # Lumi specific fix for singularity
+{% endif %}
 bash {{ aqua }}/cli/aqua-container/load-aqua-container.sh {{ machine }} <<EOF
 #!/bin/bash
 aqua lra -l info -c {{ config }} -m {{ model }} -e {{ exp }} -s {{ source }} -v {{ varname }}  -w {{ ntasks_per_node }} {{ '-d' if definitive }} {{ '-o' if overwrite }}
