+++ cli/lra/auto_LRA_overnight.yaml
@@ -4,8 +4,9 @@ target:
   frequency: monthly
   catalog: climatedt-phase1
 
-path:
-  outdir: /pfs/lustrep3/projappl/project_465000454/data/AQUA/LRA
+paths:
+  #outdir: /pfs/lustrep3/projappl/project_465000454/data/AQUA/LRA
+  outdir: /pfs/lustrep3/appl/local/climatedt/data/AQUA/LRA/
   tmpdir: /pfs/lustrep3/projappl/project_465000454/padavini/tmp
 
 
@@ -54,21 +55,21 @@ data:
         workers: 10
         vars: ['avg_thetao', 'avg_so']
 
-  IFS-NEMO:
-    control-1990:
-      hourly-hpz10-atm2d:
-        workers: 16
-        vars: ['mtpr', '2t', 'skt', 'msl', 'tcc', 'lcc', 'hcc', 'mcc', 'sd', 'tciw', 'tclw', 'mer', 'msr', 'mslhf', 'msshf',
-          'msnswrf', 'msnlwrf', 'msdwlwrf', 'msdwswrf', 'mtnswrf', 'mtnlwrf', 'mtdwswrf', 'mntss', 'metss', 'msror']
-      hourly-hpz10-atm3d:
-        workers: 6
-        vars: ['q', 't', 'u', 'v']
-      daily-hpz10-oce2d:
-        workers: 12
-        vars: ['avg_tos', 'avg_sithick', 'avg_siconc', 'avg_sos', 'avg_hc700m']
-      daily-hpz10-oce3d:
-        workers: 3
-        vars: ['avg_thetao', 'avg_so']
+  # IFS-NEMO:
+  #   control-1990:
+  #     hourly-hpz10-atm2d:
+  #       workers: 16
+  #       vars: ['mtpr', '2t', 'skt', 'msl', 'tcc', 'lcc', 'hcc', 'mcc', 'sd', 'tciw', 'tclw', 'mer', 'msr', 'mslhf', 'msshf',
+  #         'msnswrf', 'msnlwrf', 'msdwlwrf', 'msdwswrf', 'mtnswrf', 'mtnlwrf', 'mtdwswrf', 'mntss', 'metss', 'msror']
+  #     hourly-hpz10-atm3d:
+  #       workers: 6
+  #       vars: ['q', 't', 'u', 'v']
+  #     daily-hpz10-oce2d:
+  #       workers: 12
+  #       vars: ['avg_tos', 'avg_sithick', 'avg_siconc', 'avg_sos', 'avg_hc700m']
+  #     daily-hpz10-oce3d:
+  #       workers: 3
+  #       vars: ['avg_thetao', 'avg_so']
     # a1al:
     #   hourly-hpz7-atm2d:
     #     workers: 24
