+++ cli/lumi-install/environment_lumi.yml
@@ -10,6 +10,7 @@ channels:
 dependencies:
   - python>=3.9,<3.13
   - cdo>=2.2.0,<=2.4.0 #for healpix support
+  - eccodes==2.38.3
   - pandas      
   - pandoc
   - pip
